# implement optimization for long file