This would stop reading the files as soon it find a match. 

<details>

<pre>
<code>

$ time cargo run -- --debug .tmp/sample-match.txt
   Compiling cf-app-log-detector v0.1.0 (/home/olivier/workspace/cf_app_log_detector)
    Finished dev [unoptimized + debuginfo] target(s) in 1.10s
     Running `target/debug/cf-app-log-detector --debug .tmp/sample-match.txt`
[DEBUG] total number of lines: 101
[DEBUG] log lines matching: 101
[DEBUG] percentage matching: 100
.tmp/sample-match.txt is a CF application log [100% line matching]
cargo run -- --debug .tmp/sample-match.txt  0.96s user 0.32s system 107% cpu 1.188 total
 ~/workspace/cf_app_log_detector   optimization-break-on-first-line-matching ● 
$ time cargo run -- --one-line-match --debug .tmp/sample-match.txt
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
     Running `target/debug/cf-app-log-detector --one-line-match --debug .tmp/sample-match.txt`
[DEBUG] total number of lines: 1
[DEBUG] log lines matching: 1
[DEBUG] percentage matching: 100
.tmp/sample-match.txt is a CF application log [100% line matching]
cargo run -- --one-line-match --debug .tmp/sample-match.txt  0.09s user 0.02s system 98% cpu 0.109 total

</code>
</pre>
</details>